refactor: images 폴더 assets 내부로 이동
