+++ pages/_app.tsx
@@ -33,7 +33,7 @@ function MyApp({ Component, pageProps }: AppProps) {
         <ConfigProvider locale={locale}>
           <Head>
             <title>TAP SPACE</title>
-            <link rel="icon" href="/images/favicon.ico" />
+            <link rel="icon" href="/assets/images/favicon.ico" />
           </Head>
           <Script
             src="https://developers.kakao.com/sdk/js/kakao.js"
