+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
-import Mainimg from '../public/images/img1.jpg';
-import Mainimg2 from '../public/images/img2.jpg';
+import Mainimg from '../public/assets/images/img1.jpg';
+import Mainimg2 from '../public/assets/images/img2.jpg';
 import Course from '../components/home/Course';
 import Carousel from '../components/home/Carousel';
 import Slider from '../components/home/Slider';
